database-agent/pyproject.toml
[project]
name = "database-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.8.0",
    "mcp[cli]>=1.12.2",
]
